Move spawn location logic outside of the world_map
